Create README.md
SQL Server Management Studio 17.0 came out in April 25, 2017. Since then, several version and iterations have been released by Microsoft. Now the latest version of SSMS is 17.8.1.

We have also built several database servers for the past year. Every time we build a server we also install the latest copy of the SQL Server Management Studio on the server.

After a year, our servers have different versions of SQL Server Management Studio on them. We did not do a good job of keeping and upgrading our SSMS on the servers.

Since we have several environments (Development, Test, Stage and Production) it was hard to keep up with the version.